requirements.txt
django == 4.2.3
django-crispy-forms
djangorestframework==3.14.0
pytest==6.2.5
djangorestframework-simplejwt==6.0.0